requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
asyncpg==0.29.0
boto3==1.34.0
anthropic==0.28.0
groq==0.9.0
rq==1.16.1
redis==5.0.4
python-multipart==0.0.9
pydantic==2.7.0
httpx==0.27.0
